so, the ai did lots of things. i'll be testing them
