+++ brachyutils/geometry/phantom_utils.py
@@ -1707,3 +1707,141 @@ def masksToNrrd(
 
         # # Write the image
         nrrd.write(str(pth_output), all_masks, header, index_order="C")
+
+
+# Conversion utilities for phantom files
+def convert_phantom_files(
+    pth_inputs: List[Union[Path, str]],
+    type_out: str = ".nrrd",
+    dir_output: Optional[Union[Path, str]] = None,
+    multi_proc: bool = False
+) -> None:
+    """
+    Convert phantom (image and segmentation) files to the specified output format.
+    
+    Args:
+        pth_inputs: List of paths to input phantom files. Can be directories or files.
+        type_out: Output file type. Options are ".nrrd", ".dcm".
+        dir_output: Output directory path (optional).
+        multi_proc: Whether to use multiprocessing (default: False).
+    """
+    from functools import partial
+    from multiprocessing import Pool
+    from pathlib import Path
+    from tqdm import tqdm
+    
+    def _prepare_phantom_loading_item(pth_input: Path) -> dict:
+        """Prepare loading item for phantom files."""
+        base_name = pth_input.stem
+        full_suffix = "".join(pth_input.suffixes)
+        
+        if full_suffix in [".nrrd", ".nii", ".nii.gz"]:
+            # Look for matching segmentation file
+            pth_seg = pth_input.parent / f"{base_name}.seg{full_suffix}"
+            args_dict = {"pth_phantom_file": pth_input}
+            if pth_seg.exists():
+                args_dict["pth_structure_file"] = pth_seg
+        elif full_suffix in [".seg.nrrd", ".seg.nii", ".seg.nii.gz"]:
+            # Look for matching image file
+            pth_input_image = pth_input.parent / f"{base_name}{full_suffix[4:]}"
+            args_dict = {"pth_structure_file": pth_input}
+            if pth_input_image.exists():
+                args_dict["pth_phantom_file"] = pth_input_image
+            else:
+                args_dict["pth_phantom_file"] = pth_input
+        else:
+            raise ValueError(
+                f"Unsupported file type {full_suffix} for phantom conversion. "
+                "Please provide a .nrrd, .nii, .nii.gz, or a dicom directory."
+            )
+        
+        return {"loader_class": BrachyPhantom, "args_dict": args_dict}
+    
+    def _handle_dicom_directory_phantom(pth_input: Path) -> List[dict]:
+        """Process a directory containing DICOM files, return only phantom items."""
+        data_to_load = []
+        
+        if len(list(pth_input.glob("*.dcm"))) < 1:
+            print(f"No DICOM files found in the directory {pth_input}.")
+            return data_to_load
+        
+        # Handle phantom data
+        loading_phantom_item = {
+            "loader_class": BrachyPhantom,
+            "args_dict": {"dir_dicom": pth_input}
+        }
+        
+        # Check for segmentation file
+        segmentation_file = list(pth_input.glob("[Rr][Ss]*.dcm"))
+        if segmentation_file:
+            loading_phantom_item["args_dict"]["pth_structure_file"] = segmentation_file[0]
+        else:
+            print(f"No segmentation file found in the directory {pth_input}")
+        
+        data_to_load.append(loading_phantom_item)
+        return data_to_load
+    
+    def _perform_phantom_conversion(item: dict, dir_output: Path, type_out: str):
+        """Perform actual phantom conversion."""
+        loader_class = item["loader_class"]
+        args_dict = item["args_dict"]
+        
+        # Convert based on output type
+        phantom_obj = loader_class(**args_dict)
+        if type_out == ".dcm":
+            phantom_obj.export_to(dir_dicom_out=dir_output)
+        elif type_out == ".nrrd":
+            phantom_obj.export_to(dir_nrrd_out=dir_output)
+        else:
+            raise ValueError(f"Unsupported output type {type_out} for phantom conversion.")
+    
+    # Main conversion logic
+    data_to_load = []
+    
+    # Process each input path
+    for pth_input in pth_inputs:
+        pth_input = Path(pth_input)
+        if not pth_input.exists():
+            raise FileNotFoundError(f"Input file {pth_input} does not exist.")
+        
+        # Handle directories (DICOM)
+        if pth_input.is_dir():
+            dicom_data = _handle_dicom_directory_phantom(pth_input)
+            data_to_load.extend(dicom_data)
+        
+        # Handle single files
+        elif pth_input.is_file():
+            base_name = pth_input.stem
+            
+            # Skip if already processed
+            if any(
+                base_name in str(item["args_dict"].get("pth_phantom_file", ""))
+                for item in data_to_load
+            ):
+                print(f"Skipping {pth_input} as it is already in the list.")
+                continue
+                
+            data_to_load.append(_prepare_phantom_loading_item(pth_input))
+        else:
+            raise ValueError(f"Input {pth_input} is neither a file nor a directory.")
+    
+    # Check if we have valid items to process
+    if not data_to_load:
+        raise ValueError("No valid phantom files found to convert.")
+    
+    # Setup output directory
+    if dir_output is None:
+        dir_output = Path(pth_inputs[0]).parent
+    else:
+        dir_output = Path(dir_output)
+    dir_output.mkdir(parents=True, exist_ok=True)
+    
+    # Perform conversion
+    if multi_proc:
+        # Create partial function with fixed arguments
+        partial_conversion = partial(_perform_phantom_conversion, dir_output=dir_output, type_out=type_out)
+        with Pool() as pool:
+            list(tqdm(pool.imap(partial_conversion, data_to_load), total=len(data_to_load), desc="Converting phantom files"))
+    else:
+        for item in tqdm(data_to_load):
+            _perform_phantom_conversion(item, dir_output, type_out)
